Rename Utils, add tests for .equals in datafactory
