+++ src/parse.ts
@@ -6,7 +6,7 @@ import N3Parser from './n3parser'
 import { parseRDFaDOM } from './rdfaparser'
 import RDFParser from './rdfxmlparser'
 import sparqlUpdateParser from './patch-parser'
-import * as Util from './util'
+import * as Util from './utils-js'
 import Formula from './formula'
 import { ContentType } from './types'
 import { Quad } from './tf-types'
