+++ src/rdfaparser.js
@@ -16,7 +16,7 @@ import BlankNode from './blank-node'
 import Literal from './literal'
 import NamedNode from './named-node'
 import * as Uri from './uri'
-import * as Util from './util'
+import * as Util from './utils-js'
 import rdf from './factories/canonical-data-factory'
 
 if (typeof Node === 'undefined') { //  @@@@@@ Global. Interface to xmldom.
