+++ src/serializer.js
@@ -8,7 +8,7 @@
 import NamedNode from './named-node'
 import BlankNode from './blank-node'
 import * as Uri from './uri'
-import * as Util from './util'
+import * as Util from './utils-js'
 import CanonicalDataFactory from './factories/canonical-data-factory'
 import { createXSD } from './xsd'
