+++ src/update-manager.ts
@@ -11,7 +11,7 @@ import Namespace from './namespace'
 import Serializer from './serializer'
 import { join as uriJoin } from './uri'
 import { isStore, isBlankNode } from './utils/terms'
-import * as Util from './util'
+import * as Util from './utils-js'
 import Statement from './statement'
 import RDFlibNamedNode from './named-node'
 import { termValue } from './utils/termValue'
